// backend/src/constants/health.constant.d.ts
export declare enum FlowIntensity {
    LIGHT = "LIGHT",
    MEDIUM = "MEDIUM",
    HEAVY = "HEAVY",
    SPOTTING = "SPOTTING"
}
export declare enum CyclePhase {
    MENSTRUAL = "MENSTRUAL",
    FOLLICULAR = "FOLLICULAR",
    OVULATORY = "OVULATORY",
    LUTEAL = "LUTEAL"
}
export declare enum Mood {
    HAPPY = "HAPPY",
    ANXIOUS = "ANXIOUS",
    IRRITABLE = "IRRITABLE",
    CALM = "CALM",
    TIRED = "TIRED",
    ENERGETIC = "ENERGETIC",
    CRAVINGS = "CRAVINGS"
}
export declare enum PhysicalSymptom {
    CRAMPS = "CRAMPS",
    BLOATING = "BLOATING",
    HEADACHE = "HEADACHE",
    ACNE = "ACNE",
    BREAST_TENDERNESS = "BREAST_TENDERNESS",
    BACK_PAIN = "BACK_PAIN",
    NAUSEA = "NAUSEA"
}
export declare enum HealthMilestoneType {
    WATER_GOAL_STREAK = "WATER_GOAL_STREAK",
    CYCLE_LOGGED_3_MONTHS = "CYCLE_LOGGED_3_MONTHS",
    WEIGHT_TARGET_REACHED = "WEIGHT_TARGET_REACHED",
    CONSISTENT_LOGGER_7_DAYS = "CONSISTENT_LOGGER_7_DAYS",
    CYCLE_LOOGED_FIRST_TIME = "CYCLE_LOOGED_FIRST_TIME",
    WATER_LOGGED_7_DAYS = "WATER_LOGGED_7_DAYS",
    BODY_MEASURE_7_DAYS = "BODY_MEASURE_7_DAYS"
}
//# sourceMappingURL=health.constant.d.ts.map